decouple by moving dropFuc to collections/fucntions.js and change to declare all functions outside
