+++ list-operation.js
@@ -2,6 +2,8 @@
 // be executed in the renderer process for that window.
 // All of the Node.js APIs are available in this process.
 var fun = require('./import/collections/functions.js');
+const { downloadShooterSub } = require('./shooter-download-helper.js');
+
 
 window.total_download_files = 0;
 
@@ -18,9 +20,7 @@ $("#media-files-list").click(function(e){
 $("#remove-button").click(function(e){
   $(".active").remove();
   if ($(".list-group-item").length === 0){
-    //backToDragView();
     fun.backToDragView();
-    //console.log(fun.backToDrageView);
   }
 });
 
@@ -28,7 +28,6 @@ $("#download-button").click(function(e){
   $(this).text("Searching ...");
   $(this).prop('disabled', true);
 
-  const { downloadShooterSub } = require('./shooter-download-helper.js');
   let listItems = $(".list-group-item");
   total_download_files = listItems.length;
   listItems.each(function(idx, li) {
@@ -47,75 +46,6 @@ $("#download-button").click(function(e){
   // const { downloadFile } = require('./download-helper.js');
   // downloadFile('http://www.pdf995.com/samples/pdf.pdf', '/Users/haominwu/Downloads/save.pdf');
 });
-
-addMediaFile = (absolutePath) => {
-  const path = require('path');
-  $("#media-files-list").append(
-    "<li class='list-group-item'" 
-    + "absolute-path='" + absolutePath + "' "
-    + "is-downloaded='0'"
-    + ">"
-    + path.basename(absolutePath) 
-    + "</li>");
-}
-
-function getExt(filename)
-{
-    let ext = filename.split('.').pop();
-    if(ext == filename) return "";
-    return ext;
-}
-
-function isValidVideoFile(filepath){
-  const VIDEOEXTS = ['avi', 'mp4', 'mkv', 'rmvb', 'rm', 'asf', 'divx', 'mpg', 'mpeg', 'mpe', 'wmv', 'vob'];
-  let now_ext = getExt(filepath);
-  for(let i=0;i<VIDEOEXTS.length;i++){
-    if (now_ext === VIDEOEXTS[i]) return true;
-  }
-  return false;
-}
-
-function getAllFiles(root_path){
-  var fs = require('fs');
-  var res = [] , files = fs.readdirSync(root_path);
-  files.forEach(function(file){
-    var pathname = root_path + '/'+file, stat = fs.lstatSync(pathname);
-    if (!stat.isDirectory()){
-      res.push(pathname);
-    } else {
-      res = res.concat(getAllFiles(pathname));
-    }
-  });
-  return res;
-}
-
-
-function dropFuc(e) {
-  e.preventDefault();
-  
-  var items = e.dataTransfer.items;
-  for(let i=0;i<items.length;i++){
-    var entry = items[i].webkitGetAsEntry();
-    var file = items[i].getAsFile();
-    if (entry.isDirectory){
-      const filepaths = getAllFiles(file.path);
-      for (let i=0;i<filepaths.length;i++){
-        if (isValidVideoFile(filepaths[i])){
-          addMediaFile(filepaths[i]);
-        } 
-      }
-    }
-    else{
-      if (isValidVideoFile(file.path)){
-        addMediaFile(file.path);
-      }
-    }
-    if($(".list-group-item").length > 0){
-      //switchToListView();
-      fun.switchToListView();
-    }
-  }
-}
 
 window.onload = function() {
 	(function() {
@@ -148,7 +78,7 @@ window.onload = function() {
 			this.classList.remove('hover');
     }, false);
     
-    fileUploader.addEventListener('drop', dropFuc, false);
+    fileUploader.addEventListener('drop', fun.dropFuc, false);
     
     fileListUploader.addEventListener('dragover', function(e) {
 			e.preventDefault();
@@ -161,7 +91,7 @@ window.onload = function() {
 			this.classList.remove('hover');
     }, false);
     
-		fileListUploader.addEventListener('drop', dropFuc, false);
+		fileListUploader.addEventListener('drop', fun.dropFuc, false);
     
 	})();
 };
